refactor(mobile:catalog-detail): update catalog detail domain logic implementation
